modules/requirements.txt
Django==3.2.16
django-allauth==0.51.0
django-anymail==8.6
django-cors-headers==3.13.0
django-filter==22.1
django-gravatar2==1.4.4
django-timezone-field==5.0
django-sekizai==3.0.1
django-select2==7.7.3
django-tables2==2.4.1
django-short-url==1.1.6
django-phonenumber-field==7.0.0
django_redis==5.2.0
django-model-utils==4.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==5.2.0
dj-database-url==1.0.0
drf-spectacular==0.24.0
hiredis==2.0.0
mssql-django==1.1.3
phonenumbers==8.12.55
psycopg2-binary==2.9.4
pyodbc==4.0.34
pytz==2022.2.1
python-decouple==3.6
tablib==3.2.1
factory_boy
raven
